refactor: Consolidate component imports for improved organization and clarity
