+++ nextjs-logs-manager/components/index.ts
@@ -1,1 +1,3 @@
 export * from './ui/CustomNavbar';
+export * from './shared/NotAuthenticated';
+export * from './providers/Provider';
